refactor(openomni): extract message-builder from subagent runtime
